Revert "Fix"
This reverts commit 17d10abec22c71eb3c95b7a82789d72c420e4e74.

+++ vm/src/frame.rs
@@ -1,21 +1,21 @@
 pub struct Frame {
-    return_address: usize,
-    frame_pointer: usize,
+    return_address: u32,
+    frame_pointer: u32,
 }
 
 impl Frame {
-    pub fn new(return_address: usize, frame_pointer: usize) -> Self {
+    pub fn new(return_address: u32, frame_pointer: u32) -> Self {
         Self {
             return_address,
             frame_pointer,
         }
     }
 
-    pub fn return_address(&self) -> usize {
+    pub fn return_address(&self) -> u32 {
         self.return_address
     }
 
-    pub fn frame_pointer(&self) -> usize {
+    pub fn frame_pointer(&self) -> u32 {
         self.frame_pointer
     }
 }
